requirements.txt
psycopg2-binary>=2.9.9
pymysql>=1.1.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
